WLAU-Net/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pydicom>=2.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
PyWavelets>=1.1.0
tqdm>=4.62.0
PyYAML>=5.4.0
wandb>=0.12.0